Some refactor of ActionEditBox

+++ view/Aux.cpp
@@ -7,9 +7,8 @@
 namespace view {
     namespace aux {
 
-        SDL_Rect pad(const SDL_Rect& rect, size_t padding) {
-            return {rect.x - static_cast<int>(padding), rect.y - static_cast<int>(padding),
-                    static_cast<int>(rect.w + 2 * padding), static_cast<int>(rect.h + 2 * padding)};
+        SDL_Rect pad(const SDL_Rect& rect, int padding) {
+            return {rect.x - padding, rect.y - padding, rect.w + 2 * padding, rect.h + 2 * padding};
         }
 
     } // namespace aux
